- Audio -> AudioBuffer로 이름 변경 - Audio 클래스를 AudioBuffer와 AudioPlayer로 나누었음(AudioPlayer에서 여러 프로퍼티는 그대로 유지하고, 오다오 버퍼만 갈아끼울 수 있도록)

+++ Source/TGON/Core/Audio/Importer/OggVorbis/OggVorbisAudioImporter.h
@@ -185,7 +185,7 @@ inline bool OggVorbisAudioImporter<_AllocatorType>::Import(const uint8_t* srcDat
         m_channels = vorbisInfo->channels;
         m_bitsPerSample = 16; // ogg vorbis is always 16 bit.
 
-        int32_t bufferSize = ov_pcm_total(&oggVorbisFile, -1) * 2 * m_channels;
+        int32_t bufferSize = static_cast<int32_t>(ov_pcm_total(&oggVorbisFile, -1)) * 2 * static_cast<int64_t>(m_channels);
         m_soundData.resize(bufferSize);
         DecodeOggVorbis(&oggVorbisFile, &m_soundData[0], bufferSize, vorbisInfo->channels);
     }
